+++ package.json
@@ -12,7 +12,8 @@
     "fs:list": "node src/fs/list.js",
     "fs:read": "node src/fs/read.js",
     "cli:env": "set RSS_color=yellowAndBlack && set RSS_aka=school && set RSS_best=true && node src/cli/env.js",
-    "cli:args": "node src/cli/args.js"
+    "cli:args": "node src/cli/args.js",
+    "modules": "node src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "type": "git",
